Add optionlist and default form values.
Option List ensures that you can order the list of options in a picker,
and default form values ensures that changes to a form return the full
value, rather than the changed fields. This is useful for the values
of nested forms to be easily handled as it removes the need for deep
merging.